Moved formState to separate package
