+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/form/dto/FormDTO.kt
@@ -1,7 +1,7 @@
 package pl.edu.agh.ki.covid19tablet.form.dto
 
 import pl.edu.agh.ki.covid19tablet.form.FormId
-import pl.edu.agh.ki.covid19tablet.form.state.dto.FormStateDTO
+import pl.edu.agh.ki.covid19tablet.formState.dto.FormStateDTO
 import pl.edu.agh.ki.covid19tablet.schema.dto.SchemaDTO
 
 data class FormDTO(
