+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/formState/fields/dto/TextFieldStateDTO.kt
@@ -1,6 +1,6 @@
-package pl.edu.agh.ki.covid19tablet.form.state.fields.dto
+package pl.edu.agh.ki.covid19tablet.formState.fields.dto
 
-import pl.edu.agh.ki.covid19tablet.form.state.fields.TextFieldStateId
+import pl.edu.agh.ki.covid19tablet.formState.fields.TextFieldStateId
 import pl.edu.agh.ki.covid19tablet.schema.fields.TextFieldId
 
 data class TextFieldStateDTO(
