+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/schema/fields/ChoiceField.kt
@@ -1,6 +1,6 @@
 package pl.edu.agh.ki.covid19tablet.schema.fields
 
-import pl.edu.agh.ki.covid19tablet.form.state.fields.ChoiceFieldState
+import pl.edu.agh.ki.covid19tablet.formState.fields.ChoiceFieldState
 import pl.edu.agh.ki.covid19tablet.schema.fields.dto.ChoiceFieldDTO
 import javax.persistence.ElementCollection
 import javax.persistence.Entity
