+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/schema/fields/SchemaFields.kt
@@ -1,6 +1,6 @@
 package pl.edu.agh.ki.covid19tablet.schema.fields
 
-import pl.edu.agh.ki.covid19tablet.form.state.FormState
+import pl.edu.agh.ki.covid19tablet.formState.FormState
 import pl.edu.agh.ki.covid19tablet.schema.fields.dto.SchemaFieldsDTO
 import javax.persistence.Embeddable
 import javax.persistence.OneToMany
