+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/schema/fields/SignField.kt
@@ -1,6 +1,6 @@
 package pl.edu.agh.ki.covid19tablet.schema.fields
 
-import pl.edu.agh.ki.covid19tablet.form.state.fields.SignFieldState
+import pl.edu.agh.ki.covid19tablet.formState.fields.SignFieldState
 import pl.edu.agh.ki.covid19tablet.schema.fields.dto.SignFieldDTO
 import javax.persistence.Entity
 import javax.persistence.Id
