+++ backend/src/main/kotlin/pl/edu/agh/ki/covid19tablet/schema/fields/TextField.kt
@@ -1,6 +1,6 @@
 package pl.edu.agh.ki.covid19tablet.schema.fields
 
-import pl.edu.agh.ki.covid19tablet.form.state.fields.TextFieldState
+import pl.edu.agh.ki.covid19tablet.formState.fields.TextFieldState
 import pl.edu.agh.ki.covid19tablet.schema.fields.dto.TextFieldDTO
 import javax.persistence.Entity
 import javax.persistence.Id
